michaelr - GOOS, p. 93b2; trying to get a parameterized unit test for my method (getSniperCommandFromMessage) in FakeAuctionServer. JUnit's own parameterized runner sucks.
I found (what looks like) a good solution that doesn't suck in JUnitParams-1.0.1.jar. However, it suffers (what I think is) a bug. 

See
http://code.google.com/p/junitparams/issues/detail?can=2&start=0&num=100&q=&colspec=ID%20Type%20Status%20Priority%20Milestone%20Owner%20Summary&groupby=&sort=&id=46

In this commit, I saved a pretty good version FakeAuctionServerTests.java that shows how to reproduce the problem. So, I can go back here and submit this to help the junitparams author to fix the issue.

I am making some progress trying to find the root cause of the bug. To continue, I'll have to change the test in FakeAuctionServerTests.java. That is what prompted me to do this commit.