michaelr - GOOS, p. 96a; took out all of my experimentation stuff from AuctionSniperEndToEndTests.java. Now it's clean with Steps 1 - 5 right in a row.
>>>>>>>>>>>>>>> CAUTION <<<<<<<<<<<<<<<<<<<<
The book's implementation uses FakeAuctionServer as a MOCK (because it makes an assertThat() call). I think this will probably lead to a lack of readability problem when going back and having to figure out from where the failure from that assert is coming. Before I looked at the book, I returned a boolean from FakeAuctionServer.hasReceivedJoinRequestFromSniper() and did the assert on that returned value in AuctionSniperEndToEndTests. This is how I think it should be done.

I fixed a compiler error I was getting in AuctionSniperDriver.java. I now determined it is due to omission of static imports for Hamcrest Matchers.
The following method caused the compiler error:
	public void showSniperStatus(String statusText) {
		new JLabelDriver(
			this, named(Main.SNIPER_STATUS_NAME)).hasText(equalTo(statusText));
	}
This static import fixed it with unchecked warnings:
import static org.hamcrest.Matchers.*;