michaelr - GOOS, p. 125b; our class Main is "doing too many things at once" and "looks rather messy". The book "teases" out some of the behavior residing in Main into a new class AuctionSniper.
When we call the method currentPrice(), AuctionSniper will
 * 1) send a higher bid to the auction
 * 2) update the status in the user interface

The book cautions about using AuctionSniper to manipulate the UI. We should not do it. We should maintain single responsibility and SOC.

After going through the book's code, implementing myself, and getting it to run ... I see that they use a kind of interface "chaining" technique. Originally,
  chat ->| AuctionMessageTranslatr ->| Main  (see p. 118a)
         |                           |
   MessageListener           AuctionEventListener

After the chaining,
  chat ->| AuctionMessageTranslatr ->| AuctionSniper ->| Main   (see p. 125b)
         |                           |                 |
   MessageListener           AuctionEventListener   SniperListener